# packages/contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Juneflow API
  version: 0.1.0
  description: >-
    Construction ERP + Subscription SaaS (multi-tenant).
    REST, JSON, prefix /api/v1. Every request carries
    "Authorization: Bearer <jwt>"; the JWT carries company_id (tenant scope).
    Status changes go through action endpoints only, never direct PUT.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    get:
      tags: [master]
      summary: List projects (GET /x?filter&page pattern)
      operationId: listProjects
      parameters:
        - $ref: "#/components/parameters/Filter"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: >-
            Project list (tenant-scoped, B-014 paginated envelope).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Paginated"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [master]
      summary: Create project (POST /x pattern)
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created (AuditLog written by middleware).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/QuotaExceeded"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT carries company_id = tenant scope (api-contract.md).
  parameters:
    Filter:
      name: filter
      in: query
      required: false
      schema:
        type: string
      description: Free-text/structured filter (GET /x?filter&page pattern).
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
      description: 1-based page index (GET /x?filter&page pattern).
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
      description: Rows per page (B-014). Server applies a default when omitted.
...
  responses:
    Unauthorized:
      description: Missing/invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    QuotaExceeded:
      description: >-
        Package/AI quota exceeded (quota middleware per
        docs/extract/PACKAGE-RULES.md) — always code QUOTA_EXCEEDED with
        upgrade_url (PLAN.md section 5).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/QuotaExceededError"
...
  schemas:
...
    Paginated:
      type: object
      description: >-
        Standard list envelope (B-014). Every list endpoint returns this
        wrapper: data is the page of rows (item type set per endpoint via
        allOf), the rest is pagination metadata. Tenant scope still applies.
      required: [data, page, page_size, total]
      properties:
        data:
          type: array
          items: {}
        page:
          type: integer
          minimum: 1
          example: 1
        page_size:
          type: integer
          minimum: 1
          example: 20
        total:
          type: integer
          minimum: 0
          example: 137
    Error:
      type: object
      description: Standard error envelope.
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    QuotaExceededError:
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          required: [upgrade_url]
          properties:
            code:
              type: string
              enum: [QUOTA_EXCEEDED]
            upgrade_url:
              type: string
              format: uri
...
    ProjectPhase:
      type: object
      description: >-
        A phase row of a project (project_node kind=phase) — B-041(ก+).
        units = unit-kind nodes under the phase; sold_pct = round(100 ×
        sold-or-transferred sales units / units); sale_status is the phase
        node's own sale status (nullable free text).
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        units:
          type: integer
          minimum: 0
        sold_pct:
          type: integer
          minimum: 0
          maximum: 100
        sale_status:
          type: string
    Project:
      type: object
      description: >-
        Minimal placeholder per data-dictionary Project (name, type, budget,
        status). Full field modeling belongs to the schema tasks.
        short / color / company_id / units / phases are the B-041(ก+) approved
        ProjectSwitcher extensions (chrome.jsx PROJECTS): short/color are
        stamped master columns (migration 0009); company_id is the owning
        company; units / phases are derived from project_node + sales_unit.
      required: [id, name, type, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [realestate, solar, civil, service]
        budget:
          type: number
        currency_code:
          type: string
          description: Every money value carries currency_code (PLAN.md section 4).
        status:
          type: string
        short:
          type: string
        color:
          type: string
        company_id:
          type: string
          format: uuid
        units:
          type: integer
          minimum: 0
        phases:
          type: array
          items:
            $ref: "#/components/schemas/ProjectPhase"
    ProjectInput:
      type: object
      description: "Project create/update input (fields per data-dictionary)."
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [realestate, solar, civil, service]
        budget:
          type: number
        currency_code:
          type: string
        short:
          type: string
        units:
          type: integer
          minimum: 0
        phases:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              units:
                type: integer
                minimum: 0